Moved Core Repository Helper\ContentTypeDomainMapper to Mapper\ContentTypeDomainMapper
